requirements.txt
crewai
crewai-tools
python-dotenv
crawl4ai
google-generativeai>=0.3.0
httpx
fastapi
uvicorn
pydantic

